ThorinDistributions

ThorinDistributions.MultivariateGamma

ThorinDistributions.MultivariateGammaConvolution

ThorinDistributions.UnivariateGammaConvolution

Distributions.convolve

Distributions.product_distribution

ThorinDistributions.MultivariateGamma — Type

MultivariateGamma(α,θ)

Construct a distribution that correspond to several comonotonous gammas with common shapes α and respectives scales θ[i]. The distribution can then be used through several methods, following the Distributions.jl standard, mainly to obtain samples.

The code is type stable and handles any <:Real types, given by the parameters.

Note that the supprot of this distribution is a half-line in the d-dimensional space, [0,+∞], with angle given by θ.

Examples

julia> dist = MultivariateGamma(2,[3,4,5]);
julia> sample = zeros(Float64,(3,10));
julia> Random.rand!(dist,sample);

ThorinDistributions.MultivariateGammaConvolution — Type

MultivariateGammaConvolution(α,θ)

Constructs a distribution that corresponds to the convolutions of MutlivariateGamma(α[i],θ[i,:]) distributions. The distribution can then be used through several methods, following the Distributions.jl standard

The pdf and cdf and not yet coded. The code is type stable and handles any <:Real types, given by the parameters.

Fitting th edistribution is not yet possible.

Examples

julia> dist = MultivariateGammaConvolution([2,3,1],[3 0;4 0;0 1]);
julia> sample = zeros(Float64,(2,10));
julia> Random.rand!(dist,sample);

ThorinDistributions.UnivariateGammaConvolution — Type

UnivariateGammaConvolution(α,θ)

Constructs a distribution that corresponds to the convolutions of Gamma(α[i],θ[i]) distributions. The distribution can then be used through several methods, following the Distributions.jl standard, to obtain pdf, cdf, random samples...

The pdf and cdf are handled by the Moshopoulos algorithm, and random samples by simply adding random gammas. The code is type stable and handles any <:Real types, given by the parameters.

To fit the distribution, a loglikelyhood approach could be used. A more involved approach from Furman might be coded sometimes (but requires tanh-sinh integration and bigfloats...)

Examples

julia> dist = UnivariateGammaConvolution([1,0.5, 3.7],[4,2, 10]);
julia> sample = zeros(Float64,10);
julia> Random.rand!(dist,sample);
julia> pdf.((dist,),sample);

Distributions.convolve — Method

Distributions.convolve(d1,d2)

Implements the special case of covolutions of two Gamma distributions to output UnivariateGammaDistributions. Also works with d1 a gamma and d2 a UnivariateGammaConvolution, and vice versa.

Distributions.product_distribution — Method

Distributions.product_distribution(d1,d2)

Implements the special case of the product of two gamma or univariate gamma convolutions distributions, and output as the result a multivariate gamma convolution with independent margins.